% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbrosterplots-package.R
\docType{package}
\name{bbrosterplots-package}
\alias{bbrosterplots}
\alias{bbrosterplots-package}
\title{bbrosterplots: Create Clustered Heatmaps to Visualize Blood Bowl rosters}
\description{
This package allows creation of a PDF rosterbook containing all different races that participated in a tournament. For each race it makes a plot (a clustered heatmap) displaying the roster as a colorcoded matrix, where teams that are similar (based on a cost weighted similarity score) are put side by side.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gsverhoeven/bbrosterplots}
  \item Report bugs at \url{https://github.com/gsverhoeven/bbrosterplots/issues}
}

}
\author{
\strong{Maintainer}: Gertjan Verhoeven \email{gertjanverhoeven@gmail.com}

Authors:
\itemize{
  \item Hiliadan
}

}
\keyword{internal}
