% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rosterplot.R
\name{create_rosterplot}
\alias{create_rosterplot}
\title{Create rosterplot}
\usage{
create_rosterplot(
  df_rosters,
  group_name = "my_group",
  write = TRUE,
  tournament_ruleset = "my_ruleset"
)
}
\arguments{
\item{df_rosters}{A roster.}

\item{group_name}{A group name.}

\item{write}{whether to write the plot to disk or not.}

\item{tournament_ruleset}{A ruleset name.}
}
\value{
A plot object.
}
\description{
Create rosterplot
}
