% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_skill_table.R
\name{create_skill_table}
\alias{create_skill_table}
\title{Make Skill tables}
\usage{
create_skill_table(
  df_rosters,
  tournament_ruleset = "my_ruleset",
  group_name = "my_group",
  return_table = FALSE
)
}
\arguments{
\item{df_rosters}{A roster.}

\item{tournament_ruleset}{A ruleset name.}

\item{group_name}{A group name.}

\item{return_table}{If a table is returned.}
}
\value{
A tag list.
}
\description{
Make Skill tables
}
