% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_rosters}
\alias{df_rosters}
\alias{df_rosters1}
\title{FUMBBL Super League Season 6 roster data}
\format{
## `df_rosters`
A data frame with 3,573 rows and 8 columns:
\describe{
  \item{number}{number}
  \item{position}{position name}
  \item{team_id}{team_id}
}

## `df_rosters1`
A data frame with 3,521 rows and 17 columns.
}
\source{
<https://www.fumbbl.com>
}
\usage{
df_rosters

df_rosters1
}
\description{
Blood Bowl rosters from the 6th season of Super League, a competitive online
tournament where coaches are divided in divisions and can be promoted
or delegated, depending on the results.
}
\keyword{datasets}
