% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_rosterbook.R
\name{render_rosterbook}
\alias{render_rosterbook}
\title{Render rosterbook as PDF}
\usage{
render_rosterbook(target_dir = ".", params, refresh = TRUE)
}
\value{
Nothing.
}
\description{
Render rosterbook as PDF
}
