% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rosterplot.R
\name{write_rosterplot}
\alias{write_rosterplot}
\title{Write rosterplot}
\usage{
write_rosterplot(
  plot,
  plotname,
  n_legend_items,
  base_height = 6.3,
  extra_height_per_item = 0.25,
  verbose = FALSE
)
}
\arguments{
\item{plot}{A plot object}

\item{plotname}{A text string.}
}
\value{
Nothing.
}
\description{
Write rosterplot
}
