% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_roster_metadata.R
\name{add_roster_metadata}
\alias{add_roster_metadata}
\title{Add metadata to roster}
\usage{
add_roster_metadata(df_rosters, edition_val = "bb2020")
}
\arguments{
\item{df_rosters}{A (set of) roster.}

\item{edition_val}{Blood Bowl edition, either bb2020 (default) or bb2025.}
}
\value{
A set of rosters with color, cost and sort order added
}
\description{
Add metadata to roster
}
