% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rosterplot.R
\name{create_rosterplot}
\alias{create_rosterplot}
\title{Create rosterplot}
\usage{
create_rosterplot(df_rosters, group_name, race_name, write = TRUE)
}
\arguments{
\item{df_rosters}{A roster.}

\item{group_name}{A group name.}
}
\value{
A plot object.
}
\description{
Create rosterplot
}
