% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_skill_table.R
\name{make_skill_table}
\alias{make_skill_table}
\title{Make Skill tables}
\usage{
make_skill_table(race_name, df_rosters)
}
\arguments{
\item{race_name}{A race name.}

\item{df_rosters}{A roster.}
}
\value{
A tag list.
}
\description{
Make Skill tables
}
