% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_rosterbook.R
\name{render_rosterbook}
\alias{render_rosterbook}
\title{Render rosterbook as PDF}
\usage{
render_rosterbook(target_dir = "output/", params, refresh_rmd = TRUE)
}
\arguments{
\item{target_dir}{dir where the roster book is rendered.}

\item{params}{A list of parameters to customize the roster book. See example below for the available parameters.}

\item{refresh_rmd}{Controls if the template is refreshed.}
}
\value{
Nothing.
}
\description{
Render rosterbook as PDF
}
\examples{
render_rosterbook(params = list(group_name = "Super League S6",
                                tournament_ruleset = "EB25",
                                cover_image = "EB25.jpg",
                                scale_cover_perc = "100\%",
                                races = c("Khorne", "Dwarf"),
                                perc = FALSE,
                                team = FALSE), refresh_rmd = TRUE)
}
