% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_reactable.R
\name{save_reactable}
\alias{save_reactable}
\title{package uses outdated webshot, here we use webshot2}
\usage{
save_reactable(input, output, ...)
}
\description{
Note: webshot uses PhantomJS, which is a headless browser that stopped development in 2018. Please use webshot2, which uses Chromium-based browsers.
results in error: Error in tools::file_ext(output) == "png" && tools::file_ext(input) !=  :
'length = 8' in coercion to 'logical(1)': fixed using https://github.com/kcuilla/reactablefmtr/pull/59/commits/99cf8129c8975ec404a995d0d44a15d096fda603
}
